requirements.txt
PyRSMQ==0.4.5
click==8.0.3